Remove DeletedUser
The system was supposed to be a way to enforce user-bans across signups, but I didn’t like the way the feature retained personal details of deleted users.
This commit removes the feature and deletes existing data from the database.

I will find a better solution that doesn’t retain personal details :)