Automatically track logins via a plug
Currently logins via exchange tokens (remember-me cookie) do cause the login to be tracked, causing users that are using the remember-me functionality to appear like they haven’t logged in for a long time.

These changes fixes that issue by tracking logins via a unique session id that will be compared to one stored in the database for the current user.