implement incognito feature (#331)
This implements the foundation for an incognito feature.
When a user is incognito they will not be shown in the list of online users and we will not track when they logged in the last time.
Incognito users will also not be tracked in visitor log entries.

This feature is only intended for admins/mods at this point and currently no UI elements allows for switching this feature on or off. It can therefore only be activated or deactivated by changing a users incognito attribute directly in the database.